Final commit 😭, so long and thanks for all the 🐟. I mode some 🤏 changes to the code 👾, mostly micro-optimization.
For posterity's sake, SkylineSolver.java could be further improved, by
adding and optimizing the heuristics. The Documentation is also quite
complete, although some of the last additions might be a little scarce
in comments.

The Visualizer is fun to play with, run it and input a test case to look
at our brilliant solutions.

Happy packing!

- git gud

Signed-off-by: Samuel <samuel.oosterholt20001004@gmail.com>